fix: 포스트 Edit과정 수정(title로 구분하던 것 -> id로 구분)

+++ noteJ/src/main/java/com/github/jkky_98/noteJ/service/WriteService.java
@@ -41,14 +41,15 @@ public class WriteService {
         return returnSeriesList;
     }
 
-    public WriteForm getWriteEdit(User sessionUser, String postTitle) {
+    public WriteForm getWriteEdit(User sessionUser, Long postId) {
         // sessionUser <-> title 검증
         User user = userRepository.findById(sessionUser.getId()).orElseThrow(() -> new EntityNotFoundException("User 엔티티 호출 실패"));
 
         WriteForm writeForm = new WriteForm();
-        Post postEdit = postRepository.findByUserUsernameAndPostUrl(user.getUsername(), postTitle).orElseThrow(() -> new EntityNotFoundException("Post 엔티티 호출 실패"));
+        Post postEdit = postRepository.findById(postId).orElseThrow(() -> new EntityNotFoundException("Post 엔티티 호출 실패"));
 
         // tag 가져오기
+        writeForm.setId(postEdit.getId());
         writeForm.setTitle(postEdit.getTitle());
         writeForm.setTags(getTagsStringforForm(postEdit));
         writeForm.setContent(postEdit.getContent());
@@ -91,11 +92,11 @@ public class WriteService {
 
         urlProvider(form); // url 설정
 
-        String encodedContent = URLDecoder.decode(form.getContent(), StandardCharsets.UTF_8);
+        encodedContent(form);
 
         Post post = Post.builder()
                 .title(form.getTitle())
-                .content(encodedContent)
+                .content(form.getContent())
                 .writable(!form.isOpen())
                 .postSummary(form.getPostSummary())
                 .postUrl(form.getUrl())
@@ -112,8 +113,9 @@ public class WriteService {
     }
 
     private void setTag(List<Tag> tags, Post postSaved) {
+        System.out.println("##################tagRepository.saveAll(tags);##############");
         List<Tag> tagsSaved = tagRepository.saveAll(tags);
-
+        System.out.println("######################################################");
         List<PostTag> postTagsForBulkSave = new ArrayList<>();
 
         for (Tag tag : tagsSaved) {
@@ -124,22 +126,23 @@ public class WriteService {
 
             postTagsForBulkSave.add(postTag);
         }
-
+        System.out.println("###########postTagRepository.saveAll(postTagsForBulkSave);###########");
         postTagRepository.saveAll(postTagsForBulkSave);
+        System.out.println("######################################################");
     }
 
     @Transactional
-    public void saveEditWrite(WriteForm form, String title) throws IOException {
-        Post post = postRepository.findByTitle(title).orElseThrow(() -> new EntityNotFoundException("Post 엔티티 존재하지 않음"));
-        post.updatePostWithoutThumbnailAndSeries(form);
-
-        String storedFileName = null;
-        if (form.getThumbnail() != null) {
-            storedFileName = fileStore.storeFile(form.getThumbnail());
-        }
+    public void saveEditWrite(WriteForm form, Long postId) throws IOException {
+        Post post = postRepository.findById(postId).orElseThrow(() -> new EntityNotFoundException("Post not found"));
 
         Series series = seriesRepository.findBySeriesName(form.getSeries()).orElseThrow(() -> new EntityNotFoundException("Series 엔티티 존재하지 않음"));
         post.updateSeries(series);
+
+        urlProvider(form); // url 설정
+
+        encodedContent(form); // content 설정
+
+        post.updatePostWithoutThumbnailAndSeries(form);
 
         // 기존 Post-PostTag 관계 제거
         for (PostTag postTag : post.getPostTags()) {
@@ -147,15 +150,26 @@ public class WriteService {
         }
         post.getPostTags().clear(); // Post와의 관계 끊기
 
-        String thumbnailDeleted = post.getThumbnail();
-        fileStore.storeFile(form.getThumbnail()); // 새로운 파일 업로드
-        post.updateThumbnail(storedFileName);
-        fileStore.deleteFile(thumbnailDeleted); // 기존 파일 삭제
+        String storedFileName = null;
+        if (form.getThumbnail() != null && !form.getThumbnail().isEmpty()) {
+            String thumbnailDeleted = post.getThumbnail();
+            storedFileName = fileStore.storeFile(form.getThumbnail());// 새로운 파일 업로드
+            // 기존 파일이 존재할 경우
+            if (post.getThumbnail() != null) {
+                post.updateThumbnail(storedFileName);
+                fileStore.deleteFile(thumbnailDeleted); // 기존 파일 삭제
+            }
+        }
 
         //Tag 업데이트
         List<Tag> tags = tagProvider(form.getTags());
         setTag(tags, post);
 
+    }
+
+    private static void encodedContent(WriteForm form) {
+        String encodedContent = URLDecoder.decode(form.getContent(), StandardCharsets.UTF_8); // content 설정
+        form.setContent(encodedContent);
     }
